# reports/rag_ingestion_auth_review/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromptPreviewResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "vector",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "prompt_context_items": {
            "items": {
              "$ref": "#/components/schemas/RetrievalResultItem"
            },
            "title": "Prompt Context Items",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RetrievalResultItem"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "mode",
          "count",
          "prompt",
          "prompt_context_items",
          "results"
        ],
        "title": "PromptPreviewResponse",
        "type": "object"
      },
...
      "RagConfigResponse": {
        "properties": {
          "embedding_model_name": {
            "title": "Embedding Model Name",
            "type": "string"
          },
          "faiss_docs_path": {
            "title": "Faiss Docs Path",
            "type": "string"
          },
          "faiss_index_path": {
            "title": "Faiss Index Path",
            "type": "string"
          },
          "faiss_store_dir": {
            "title": "Faiss Store Dir",
            "type": "string"
          },
          "min_vector_score": {
            "title": "Min Vector Score",
            "type": "number"
          },
          "model_rerank_weight": {
            "title": "Model Rerank Weight",
            "type": "number"
          },
          "reply_rules_enabled": {
            "title": "Reply Rules Enabled",
            "type": "boolean"
          },
          "reranker_model_name": {
            "title": "Reranker Model Name",
            "type": "string"
          }
        },
        "required": [
          "embedding_model_name",
          "reranker_model_name",
          "model_rerank_weight",
          "min_vector_score",
          "faiss_store_dir",
          "faiss_index_path",
          "faiss_docs_path",
          "reply_rules_enabled"
        ],
        "title": "RagConfigResponse",
        "type": "object"
      },
...
      "RetrievalResultItem": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "direction_penalty": {
            "title": "Direction Penalty",
            "type": "number"
          },
          "display_title": {
            "default": "",
            "title": "Display Title",
            "type": "string"
          },
          "evidence_strength": {
            "default": "normal",
            "title": "Evidence Strength",
            "type": "string"
          },
          "evidence_summary": {
            "default": "",
            "title": "Evidence Summary",
            "type": "string"
          },
          "intent": {
            "title": "Intent",
            "type": "string"
          },
          "keyword_bonus": {
            "title": "Keyword Bonus",
            "type": "number"
          },
          "model_rerank_score": {
            "title": "Model Rerank Score",
            "type": "number"
          },
          "prompt_instruction": {
            "default": "",
            "title": "Prompt Instruction",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "rerank_score": {
            "title": "Rerank Score",
            "type": "number"
          },
          "role": {
            "default": "",
            "title": "Role",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_answer": {
            "default": "",
            "title": "Source Answer",
            "type": "string"
          },
          "source_question": {
            "default": "",
            "title": "Source Question",
            "type": "string"
          },
          "vector_score": {
            "title": "Vector Score",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "score",
          "rerank_score",
          "model_rerank_score",
          "vector_score",
          "keyword_bonus",
          "direction_penalty",
          "category",
          "intent",
          "question",
          "answer"
        ],
        "title": "RetrievalResultItem",
        "type": "object"
      },
      "RetrievalSearchRequest": {
        "properties": {
          "limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "min_score": {
            "default": 0.4,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Score",
            "type": "number"
          },
          "mode": {
            "default": "hybrid",
            "enum": [
              "vector",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "RetrievalSearchRequest",
        "type": "object"
      },
      "RetrievalSearchResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "vector",
              "hybrid"
            ],
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RetrievalResultItem"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "mode",
          "count",
          "results"
        ],
        "title": "RetrievalSearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "身份只来自 Authorization: Bearer <JWT>。X-User-Role / X-Operator-Id 已失效，仅在日志中记录，不参与任何授权判定。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LLM Customer Service API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/retrieval/config": {
      "get": {
        "operationId": "get_retrieval_config_retrieval_config_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Retrieval Config",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/retrieval/prompt-preview": {
      "post": {
        "operationId": "preview_prompt_retrieval_prompt_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Prompt",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/retrieval/search": {
      "post": {
        "operationId": "search_retrieval_retrieval_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Operator-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Retrieval",
        "tags": [
          "retrieval"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}